+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/handlers/JsonRpcExecutorHandler.java
@@ -17,6 +17,7 @@ package org.hyperledger.besu.ethereum.api.handlers;
 import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError.INVALID_REQUEST;
 
 import org.hyperledger.besu.ethereum.api.jsonrpc.JsonResponseStreamer;
+import org.hyperledger.besu.ethereum.api.jsonrpc.JsonRpcConfiguration;
 import org.hyperledger.besu.ethereum.api.jsonrpc.context.ContextKey;
 import org.hyperledger.besu.ethereum.api.jsonrpc.execution.JsonRpcExecutor;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.JsonRpcRequest;
@@ -26,6 +27,7 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcRespon
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponseType;
 
 import java.io.IOException;
+import java.security.InvalidParameterException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -64,71 +66,27 @@ public class JsonRpcExecutorHandler {
   private JsonRpcExecutorHandler() {}
 
   public static Handler<RoutingContext> handler(
-      final JsonRpcExecutor jsonRpcExecutor, final Tracer tracer) {
+      final JsonRpcExecutor jsonRpcExecutor,
+      final Tracer tracer,
+      final JsonRpcConfiguration jsonRpcConfiguration) {
     return ctx -> {
       HttpServerResponse response = ctx.response();
       try {
-        Optional<User> user = ContextKey.AUTHENTICATED_USER.extractFrom(ctx, Optional::empty);
-        Context spanContext = ctx.get(SPAN_CONTEXT);
         response = response.putHeader("Content-Type", APPLICATION_JSON);
-
-        if (ctx.data().containsKey(ContextKey.REQUEST_BODY_AS_JSON_OBJECT.name())) {
-          JsonObject jsonRequest = ctx.get(ContextKey.REQUEST_BODY_AS_JSON_OBJECT.name());
-          lazyTraceLogger(jsonRequest::toString);
-          JsonRpcResponse jsonRpcResponse =
-              jsonRpcExecutor.execute(
-                  user,
-                  tracer,
-                  spanContext,
-                  () -> !ctx.response().closed(),
-                  jsonRequest,
-                  req -> req.mapTo(JsonRpcRequest.class));
-          response.setStatusCode(status(jsonRpcResponse).code());
-          if (jsonRpcResponse.getType() == JsonRpcResponseType.NONE) {
-            response.end();
-          } else {
-            try (final JsonResponseStreamer streamer =
-                new JsonResponseStreamer(response, ctx.request().remoteAddress())) {
-              // underlying output stream lifecycle is managed by the json object writer
-              lazyTraceLogger(() -> JSON_OBJECT_MAPPER.writeValueAsString(jsonRpcResponse));
-              JSON_OBJECT_WRITER.writeValue(streamer, jsonRpcResponse);
-            }
-          }
-        } else if (ctx.data().containsKey(ContextKey.REQUEST_BODY_AS_JSON_ARRAY.name())) {
-          JsonArray batchJsonRequest = ctx.get(ContextKey.REQUEST_BODY_AS_JSON_ARRAY.name());
-          lazyTraceLogger(batchJsonRequest::toString);
-          List<JsonRpcResponse> jsonRpcBatchResponses = new ArrayList<>();
+        if (isJsonObjectRequest(ctx)) {
+          final JsonRpcResponse jsonRpcResponse =
+              executeJsonObjectRequest(jsonRpcExecutor, tracer, ctx);
+          handleJsonObjectResponse(response, jsonRpcResponse, ctx);
+        } else if (isJsonArrayRequest(ctx)) {
+          final List<JsonRpcResponse> jsonRpcBatchResponses;
           try {
-            for (int i = 0; i < batchJsonRequest.size(); i++) {
-              final JsonObject jsonRequest;
-              try {
-                jsonRequest = batchJsonRequest.getJsonObject(i);
-              } catch (ClassCastException e) {
-                jsonRpcBatchResponses.add(new JsonRpcErrorResponse(null, INVALID_REQUEST));
-                continue;
-              }
-              jsonRpcBatchResponses.add(
-                  jsonRpcExecutor.execute(
-                      user,
-                      tracer,
-                      spanContext,
-                      () -> !ctx.response().closed(),
-                      jsonRequest,
-                      req -> req.mapTo(JsonRpcRequest.class)));
-            }
-          } catch (RuntimeException e) {
+            jsonRpcBatchResponses =
+                executeJsonArrayRequest(jsonRpcExecutor, tracer, ctx, jsonRpcConfiguration);
+            handleJsonArrayResponse(response, jsonRpcBatchResponses, ctx);
+          } catch (final InvalidParameterException e) {
+            handleJsonRpcError(ctx, null, JsonRpcError.EXCEEDS_RPC_MAX_BATCH_SIZE);
+          } catch (final RuntimeException e) {
             response.setStatusCode(HttpResponseStatus.BAD_REQUEST.code()).end();
-            return;
-          }
-          final JsonRpcResponse[] completed =
-              jsonRpcBatchResponses.stream()
-                  .filter(jsonRpcResponse -> jsonRpcResponse.getType() != JsonRpcResponseType.NONE)
-                  .toArray(JsonRpcResponse[]::new);
-          try (final JsonResponseStreamer streamer =
-              new JsonResponseStreamer(response, ctx.request().remoteAddress())) {
-            // underlying output stream lifecycle is managed by the json object writer
-            lazyTraceLogger(() -> JSON_OBJECT_MAPPER.writeValueAsString(completed));
-            JSON_OBJECT_WRITER.writeValue(streamer, completed);
           }
         } else {
           handleJsonRpcError(ctx, null, JsonRpcError.PARSE_ERROR);
@@ -142,8 +100,102 @@ public class JsonRpcExecutorHandler {
     };
   }
 
+  private static boolean isJsonObjectRequest(final RoutingContext ctx) {
+    return ctx.data().containsKey(ContextKey.REQUEST_BODY_AS_JSON_OBJECT.name());
+  }
+
+  private static boolean isJsonArrayRequest(final RoutingContext ctx) {
+    return ctx.data().containsKey(ContextKey.REQUEST_BODY_AS_JSON_ARRAY.name());
+  }
+
+  private static JsonRpcResponse executeRequest(
+      final JsonRpcExecutor jsonRpcExecutor,
+      final Tracer tracer,
+      final JsonObject jsonRequest,
+      final RoutingContext ctx) {
+    final Optional<User> user = ContextKey.AUTHENTICATED_USER.extractFrom(ctx, Optional::empty);
+    final Context spanContext = ctx.get(SPAN_CONTEXT);
+    return jsonRpcExecutor.execute(
+        user,
+        tracer,
+        spanContext,
+        () -> !ctx.response().closed(),
+        jsonRequest,
+        req -> req.mapTo(JsonRpcRequest.class));
+  }
+
+  private static JsonRpcResponse executeJsonObjectRequest(
+      final JsonRpcExecutor jsonRpcExecutor, final Tracer tracer, final RoutingContext ctx) {
+    final JsonObject jsonRequest = ctx.get(ContextKey.REQUEST_BODY_AS_JSON_OBJECT.name());
+    lazyTraceLogger(jsonRequest::toString);
+    return executeRequest(jsonRpcExecutor, tracer, jsonRequest, ctx);
+  }
+
+  private static List<JsonRpcResponse> executeJsonArrayRequest(
+      final JsonRpcExecutor jsonRpcExecutor,
+      final Tracer tracer,
+      final RoutingContext ctx,
+      final JsonRpcConfiguration jsonRpcConfiguration)
+      throws InvalidParameterException {
+    final JsonArray batchJsonRequest = ctx.get(ContextKey.REQUEST_BODY_AS_JSON_ARRAY.name());
+    lazyTraceLogger(batchJsonRequest::toString);
+    final List<JsonRpcResponse> jsonRpcBatchResponses = new ArrayList<>();
+
+    if (jsonRpcConfiguration.getMaxBatchSize() > 0
+        && batchJsonRequest.size() > jsonRpcConfiguration.getMaxBatchSize()) {
+      throw new InvalidParameterException();
+    }
+
+    for (int i = 0; i < batchJsonRequest.size(); i++) {
+      final JsonObject jsonRequest;
+      try {
+        jsonRequest = batchJsonRequest.getJsonObject(i);
+      } catch (final ClassCastException e) {
+        jsonRpcBatchResponses.add(new JsonRpcErrorResponse(null, INVALID_REQUEST));
+        continue;
+      }
+      jsonRpcBatchResponses.add(executeRequest(jsonRpcExecutor, tracer, jsonRequest, ctx));
+    }
+    return jsonRpcBatchResponses;
+  }
+
+  private static void handleJsonObjectResponse(
+      final HttpServerResponse response,
+      final JsonRpcResponse jsonRpcResponse,
+      final RoutingContext ctx)
+      throws IOException {
+    response.setStatusCode(status(jsonRpcResponse).code());
+    if (jsonRpcResponse.getType() == JsonRpcResponseType.NONE) {
+      response.end();
+    } else {
+      try (final JsonResponseStreamer streamer =
+          new JsonResponseStreamer(response, ctx.request().remoteAddress())) {
+        // underlying output stream lifecycle is managed by the json object writer
+        lazyTraceLogger(() -> JSON_OBJECT_MAPPER.writeValueAsString(jsonRpcResponse));
+        JSON_OBJECT_WRITER.writeValue(streamer, jsonRpcResponse);
+      }
+    }
+  }
+
+  private static void handleJsonArrayResponse(
+      final HttpServerResponse response,
+      final List<JsonRpcResponse> jsonRpcBatchResponses,
+      final RoutingContext ctx)
+      throws IOException {
+    final JsonRpcResponse[] completed =
+        jsonRpcBatchResponses.stream()
+            .filter(jsonRpcResponse -> jsonRpcResponse.getType() != JsonRpcResponseType.NONE)
+            .toArray(JsonRpcResponse[]::new);
+    try (final JsonResponseStreamer streamer =
+        new JsonResponseStreamer(response, ctx.request().remoteAddress())) {
+      // underlying output stream lifecycle is managed by the json object writer
+      lazyTraceLogger(() -> JSON_OBJECT_MAPPER.writeValueAsString(completed));
+      JSON_OBJECT_WRITER.writeValue(streamer, completed);
+    }
+  }
+
   private static String getRpcMethodName(final RoutingContext ctx) {
-    if (ctx.data().containsKey(ContextKey.REQUEST_BODY_AS_JSON_OBJECT.name())) {
+    if (isJsonObjectRequest(ctx)) {
       final JsonObject jsonObject = ctx.get(ContextKey.REQUEST_BODY_AS_JSON_OBJECT.name());
       return jsonObject.getString("method");
     } else {