+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/MainnetBlockValidator.java
@@ -186,13 +186,17 @@ public class MainnetBlockValidator implements BlockValidator {
       final Block invalidBlock, final BlockValidationResult result) {
     if (result.causedBy().isPresent()) {
       LOG.info(
-          "{}. Block {}, caused by {}",
-          result.errorMessage,
+          "Invalid block {}: {}, caused by {}",
           invalidBlock.toLogString(),
+          result.errorMessage,
           result.causedBy().get());
       LOG.debug("with stack", result.causedBy().get());
     } else {
-      LOG.info("{}. Block {}", result.errorMessage, invalidBlock.toLogString());
+      if (result.errorMessage.isPresent()) {
+        LOG.info("Invalid block {}: {}", invalidBlock.toLogString(), result.errorMessage);
+      } else {
+        LOG.info("Invalid block {}", invalidBlock.toLogString());
+      }
     }
     badBlockManager.addBadBlock(invalidBlock, result.causedBy());
   }