+++ evm/src/main/java/org/hyperledger/besu/evm/operation/DataHashOperation.java
@@ -24,10 +24,23 @@ import java.util.Optional;
 
 import org.apache.tuweni.bytes.Bytes;
 
+/**
+ * The DataHash operation. https://eips.ethereum.org/EIPS/eip-4844
+ *
+ * <p>Reads index from the top of the stack as big-endian uint256, and replaces it on the stack with
+ * tx.message.blob_versioned_hashes[index] if index &lt; len(tx.message.blob_versioned_hashes), and
+ * otherwise with a zeroed bytes32 value.
+ */
 public class DataHashOperation extends AbstractOperation {
 
+  /** DATAHASH opcode number */
   public static final int OPCODE = 0x49;
 
+  /**
+   * Instantiates a new DataHash operation.
+   *
+   * @param gasCalculator the gas calculator
+   */
   public DataHashOperation(final GasCalculator gasCalculator) {
     super(OPCODE, "DATAHASH", 1, 1, gasCalculator);
   }